+++ .github/actions/publish-action/entrypoint.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+make clean deps lint package
+echo "${GITHUB_TOKEN}" | gh auth login --with-token
+make publish