+++ .github/actions/release-minor-action/entrypoint.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+git config --global user.email "shineworks@shinesolutions.com"
+git config --global user.name "Shine Works"
+make release-minor