+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Option to execute an Apache graceful or hard restart when deploying artifacts
 - Set Python3 to be the default for Ansible
 - Proxy provisioning feature toggle on cloud-init user data for both Full-Set and Consolidated
+- Add release-major, release-minor, release-patch, and publish Makefile targets and GitHub Actions
 
 ### Fixed
 - Fix CloudWatch Dashboard load balancer metrics #435